[cmake] avoid use of undocumented FindThread module variable
df7103ddc555 ([build] Make sure to link main executable with pthreads,
2016-06-21) add a reference to an internal variable that was meant to
ensure the returned flag would be a library name.

Change the variable name to the documented[1] one and that was made
available since release of cmake 3.1.

After this change; switching its value to "true" and removing the TODO
should be possible when using newer cmake, and indeed seems to work
fine under limited testing, but that change has been punted until a
need arises (ex: RISCV with gcc failing to link because of missing
libatomic) since it could also be problematic (ex: older clang warning
when it is used while linking)

[1] https://cmake.org/cmake/help/latest/module/FindThreads.html#variable:THREADS_PREFER_PTHREAD_FLAG

Differential Revision: https://reviews.llvm.org/D122932